Allow user to specify document format
Currently, we are responding with both `json` and `xml` document in the
response, which makes our payload much larger that in need to be. In a
real world scenario, a user will need one or the either, so this commit
adds support for `document_format` params in the requests. Based on this
the user will either receive `json` and `xml` doc.