Adopt relaton to parse and fetch standards
The current relaton library is doing most of the work we need to do
in this application, like parsing the code, and then fetching the
actual document from a specific standard body. At the same time we
also want this service to be used by relaton tool chain, so we also
need to make sure we are not creating any circular dependency.

So, initially we are starting to use the relaton library as it is
and later when we are ready, then we might consider separating the
relaton core standard fetching and db maintaining functionality to
separate libraries.