Improve docker-compose setup for development
Current docker file has two issues, one is it tries to install
bundler dependencies in each build and second of all postgres
data resets in each setup. This adds some extra time when we
need to do a quick test.

This commit changes these, and it adds docker's mountable
volumes, so now whenever we run bundler then it will use this
directory to store it's dependencies and also share between
these builds. Same, goes for the postgres data persistence.